Improve diff highlighting
DiifChange, DiffAdd and DiffDelete colors are now higher contrast with
the default background. I was trying to make them subtle before, but
they were too subtle.

I'd also tried using undercurl for DiffText as an experiment, but simply
could not get used to it so I've now switched it to use to a more
intense version of the DiffChange background color. (This is similar to
the viewsource colorscheme's approach, though geared towards a dark
background. Also, unlike viewsource, holodark preserves the foreground
coloration and styling of added/changed text, which does make finding
the right colors a bit more challenging.)